[ENH] parallelization backends for grid and random search (#162)
From Original PR #150:

This PR introduces parallelization backends for GridSearch and
RandomSearch.

Both estimators now allow to specify backend and backend_params in the
constructor, which allows selection of a parallelization backend and
configuration parameters for it - the default being "None", i.e., plain
loop.

This uses the parallel utilities also used in sktime and skpro, with a
mid-term plan to move these to scikit-base.

---------

Co-authored-by: Franz Király <fkiraly@gcos.ai>