Add PyTorch Lightning integration (#203)
This pull request is to merge a new experiment
`torch-lightning-experiment` for the issue `195`. The new class
`TorchExperiment` is an experiment adapter for PyTorch Lightning
experiments and is used to perform experiments using PyTorch Lightning
modules. It allows for hyperparameter tuning and evaluation of the
model's performance using specified metrics.

The `TorchExperiment` class accepts a `LightningModule`, `DataModule`,
`Trainer`, and an `objective_matric` with default value `val_loss`. The
`_evaluate` function internally performs a training run and returns a
score of the `objective_matric`.

Fixes #195